s3s: checksum: ChecksumHasher
