+++ crates/s3s-fs/Cargo.toml
@@ -26,11 +26,8 @@ chrono = { version = "0.4.26", default-features = false, features = [
 ] }
 clap = { version = "4.3.21", optional = true, features = ["derive"] }
 crc32c = "0.6.4"
-crc32fast = "1.3.2"
-digest = "0.10.7"
 futures = "0.3.28"
 hex-simd = "0.8.0"
-
 hyper-util = { version = "0.1.3", optional = true, features = [
     "server",
     "http1",
@@ -44,8 +41,6 @@ numeric_cast = "0.2.1"
 path-absolutize = "3.1.0"
 s3s = { version = "0.10.1-dev", path = "../s3s" }
 serde_json = "1.0.104"
-sha1 = "0.10.5"
-sha2 = "0.10.7"
 thiserror = "1.0.45"
 time = "0.3.25"
 tokio = { version = "1.31.0", features = ["fs", "io-util"] }
