+++ crates/s3s/Cargo.toml
@@ -27,7 +27,9 @@ base64-simd = "0.8.0"
 bytes = "1.4.0"
 bytestring = "1.3.0"
 chrono = { version = "0.4.26", default-features = false }
+crc32c = "0.6.4"
 crc32fast = "1.3.2"
+digest = "0.10.7"
 futures = { version = "0.3.28", default-features = false, features = ["std"] }
 hex-simd = "0.8.0"
 hmac = "0.12.1"
@@ -40,6 +42,7 @@ memchr = "2.6.2"
 mime = "0.3.17"
 nom = "7.1.3"
 nugine-rust-utils = "0.3.1"
+numeric_cast = "0.2.1"
 pin-project-lite = "0.2.12"
 quick-xml = { version = "0.31.0", features = ["serialize"] }
 serde = { version = "1.0.183", features = ["derive"] }
