+++ crates/s3s/src/lib.rs
@@ -42,6 +42,8 @@ pub mod path;
 pub mod service;
 pub mod stream;
 
+pub mod checksum;
+
 pub use self::error::*;
 pub use self::http::Body;
 pub use self::request::S3Request;
